+++ packages/amplify-migration/src/app_storage_definition_fetcher.ts
@@ -4,18 +4,27 @@ import path from 'node:path';
 import { getStorageDefinition } from '@aws-amplify/amplify-gen1-codegen-storage-adapter';
 import { BackendDownloader } from './backend_downloader.js';
 import { fileOrDirectoryExists } from './directory_exists.js';
+import { StorageRenderParameters } from '@aws-amplify/amplify-gen2-codegen';
+import { GetBucketNotificationConfigurationCommand, S3Client } from '@aws-sdk/client-s3';
+import { BackendEnvironmentResolver } from './backend_environment_selector';
 
 export interface AppStorageDefinitionFetcher {
-  getDefinition(deploymentBucket: string): Promise<ReturnType<typeof getStorageDefinition> | undefined>;
+  getDefinition(): Promise<ReturnType<typeof getStorageDefinition> | undefined>;
 }
 export class AppStorageDefinitionFetcher {
-  constructor(private ccbFetcher: BackendDownloader) {}
+  constructor(
+    private backendEnvironmentResolver: BackendEnvironmentResolver,
+    private ccbFetcher: BackendDownloader,
+    private s3Client: S3Client,
+  ) {}
   private readJsonFile = async (filePath: string) => {
     const contents = await fs.readFile(filePath, { encoding: 'utf8' });
     return JSON.parse(contents);
   };
-  getDefinition = async (deploymentBucket: string) => {
-    const currentCloudBackendDirectory = await this.ccbFetcher.getCurrentCloudBackend(deploymentBucket);
+  getDefinition = async (): Promise<StorageRenderParameters | undefined> => {
+    const backendEnvironment = await this.backendEnvironmentResolver.selectBackendEnvironment();
+    assert(backendEnvironment?.deploymentArtifacts);
+    const currentCloudBackendDirectory = await this.ccbFetcher.getCurrentCloudBackend(backendEnvironment.deploymentArtifacts);
 
     const amplifyMetaPath = path.join(currentCloudBackendDirectory, 'amplify-meta.json');
 
@@ -27,9 +36,14 @@ export class AppStorageDefinitionFetcher {
       const cliInputsPath = path.join(currentCloudBackendDirectory, 'storage', storageName, 'cli-inputs.json');
       assert(await fileOrDirectoryExists(cliInputsPath));
       const cliInputs = await this.readJsonFile(cliInputsPath);
+      assert(cliInputs.bucketName);
+      const { bucketName } = cliInputs;
+      console.log(bucketName);
+      const triggers = await this.s3Client.send(new GetBucketNotificationConfigurationCommand({ Bucket: bucketName }));
+      console.log('triggers', triggers);
       const storageOptions = getStorageDefinition({
         cliInputs,
-        bucketName: cliInputs['bucketName'],
+        bucketName,
       });
       return storageOptions;
     }