+++ packages/amplify-util-mock/src/__e2e_v2__/test-synthesizer/transform.ts
@@ -22,6 +22,7 @@ export const defaultTransformParams: Pick<ExecuteTransformConfig, 'transformersF
     allowDestructiveGraphqlSchemaUpdates: false,
     replaceTableUponGsiUpdate: false,
     subscriptionsInheritPrimaryAuth: false,
+    allowGen1Patterns: true,
   },
 };