+++ packages/amplify-appsync-simulator/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.16.5](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-appsync-simulator@2.16.4...@aws-amplify/amplify-appsync-simulator@2.16.5) (2024-08-02)
+
+**Note:** Version bump only for package @aws-amplify/amplify-appsync-simulator
+
+
+
+
+
 ## [2.16.4](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-appsync-simulator@2.16.3...@aws-amplify/amplify-appsync-simulator@2.16.4) (2024-06-27)
 
 **Note:** Version bump only for package @aws-amplify/amplify-appsync-simulator